isolate aircraft specific logic
